+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 
+## 1.2.0
+
+- Added separator argument (`--`)
+- Added option to set AWS Profile via additional arguments
+
 ## 1.1.0
 
-Added additional arguments
+- Added additional arguments
 
 ## 1.0.0
 
-Added beta code of the tool to support Linux and Windows OS.
+- Added beta code of the tool to support Linux and Windows OS.